Changed db scheme, added notes feature, added multiples of maharaj and guruji images.

+++ app/src/main/java/com/swaminarayan/shikshapatriApp/domain/models/ReportAgnaItem.kt
@@ -3,6 +3,6 @@ package com.swaminarayan.shikshapatriApp.domain.models
 data class ReportAgnaItem(
     val agnaId: Long,
     val agna: String,
-    val totalAgnaPoints: Int,
-    val agnaPalaiPoints: Int
+    val totalAgnaPoints: Long,
+    val agnaPalaiPoints: Long
 )
